Move lsn_to_bytes and bytes_to_lsn into Postgresql class
and make their behavior version specific
